+++ src/main/java/com/icomfortableworld/global/exception/GlobalExceptionHandler.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import com.icomfortableworld.common.dto.CommonResponseDto;
-import com.icomfortableworld.global.exception.entity.CustomMemberException;
 import com.icomfortableworld.global.exception.jwt.CustomJwtException;
+import com.icomfortableworld.global.exception.member.CustomMemberException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
