+++ src/main/java/com/icomfortableworld/global/exception/member/MemberErrorCode.java
@@ -1,4 +1,4 @@
-package com.icomfortableworld.global.exception.entity;
+package com.icomfortableworld.global.exception.member;
 
 import lombok.Getter;
